# src/main/resources/openapi.yml
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenApi specification - Ousmane Vincent Dione",
    "description": "OpenApi documentation for Stock Management",
    "termsOfService": "Terms of service",
    "contact": {
      "name": "Ousmane Vincent",
      "url": "https://ndiolfouta.com",
      "email": "dioneousmanevincent@gmail.com"
    },
    "license": {
      "name": "Licence name",
      "url": "https://some-url.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8082",
      "description": "Local ENV"
    },
    {
      "url": "https://ndiolfouta.com",
      "description": "PROD ENV"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/clients": {
      "post": {
        "tags": [
          "client-api"
        ],
        "operationId": "save_6",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/categories": {
      "post": {
        "tags": [
          "category-api"
        ],
        "operationId": "save_7",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "authentication-controller"
        ],
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/authenticate": {
      "post": {
        "tags": [
          "authentication-controller"
        ],
        "operationId": "authenticate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthenticationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/articles": {
      "post": {
        "tags": [
          "article-api"
        ],
        "operationId": "save_8",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/clients/{idClient}": {
      "get": {
        "tags": [
          "client-api"
        ],
        "operationId": "getClientById",
        "parameters": [
          {
            "name": "idClient",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/categories/{nom}": {
      "get": {
        "tags": [
          "category-api"
        ],
        "operationId": "getCategoryByNom",
        "parameters": [
          {
            "name": "nom",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/categories/{idCategory}": {
      "get": {
        "tags": [
          "category-api"
        ],
        "operationId": "getCategoryById",
        "parameters": [
          {
            "name": "idCategory",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/articles/{idArticle}": {
      "get": {
        "tags": [
          "article-api"
        ],
        "operationId": "getArticleById",
        "parameters": [
          {
            "name": "idArticle",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/articles/historique/ventes/{idArticle}": {
      "get": {
        "tags": [
          "article-api"
        ],
        "operationId": "findHistoriqueVente",
        "parameters": [
          {
            "name": "idArticle",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LigneVenteDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/articles/historique/commandeFournisseur/{idArticle}": {
      "get": {
        "tags": [
          "article-api"
        ],
        "operationId": "findHistoriqueCommandeFournisseur",
        "parameters": [
          {
            "name": "idArticle",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LigneCommandeFournisseurDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/articles/historique/commandeClients/{idArticle}": {
      "get": {
        "tags": [
          "article-api"
        ],
        "operationId": "findHistoriqueCommandeClient",
        "parameters": [
          {
            "name": "idArticle",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LigneCommandeClientDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/articles/filter/category/{idCategory}": {
      "get": {
        "tags": [
          "article-api"
        ],
        "operationId": "findHistoriqueByIdCatergory",
        "parameters": [
          {
            "name": "idCategory",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArticleDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/commandeClients/deleteArticle/{idCommande}/{idLigneCommande}": {
      "delete": {
        "tags": [
          "commande-client-api"
        ],
        "operationId": "deleteArticle",
        "parameters": [
          {
            "name": "idCommande",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "idLigneCommande",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommandeClientDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/clients/delete/{idClient}": {
      "delete": {
        "tags": [
          "client-api"
        ],
        "operationId": "deleteClient",
        "parameters": [
          {
            "name": "idClient",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/categories/delete/{idCategory}": {
      "delete": {
        "tags": [
          "category-api"
        ],
        "operationId": "deleteCategory",
        "parameters": [
          {
            "name": "idCategory",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "VenteDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string"
          },
          "dateVente": {
            "type": "string",
            "format": "date-time"
          },
          "commentaire": {
            "type": "string"
          }
        }
      },
      "AdresseDto": {
        "type": "object",
        "properties": {
          "adresse1": {
            "type": "string"
          },
          "adresse2": {
            "type": "string"
          },
          "ville": {
            "type": "string"
          },
          "codePostal": {
            "type": "string"
          },
          "pays": {
            "type": "string"
          }
        }
      },
...
      "CommandeFournisseurDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string"
          },
          "dateCommande": {
            "type": "string",
            "format": "date-time"
          },
          "fournisseurDto": {
            "$ref": "#/components/schemas/FournisseurDto"
          }
        }
      },
      "FournisseurDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nom": {
            "type": "string"
          },
          "prenom": {
            "type": "string"
          },
          "adresseDto": {
            "$ref": "#/components/schemas/AdresseDto"
          },
          "photo": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "numTel": {
            "type": "string"
          },
          "commandeFournisseurDtos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommandeFournisseurDto"
            }
          }
        }
      },
      "Adresse": {
        "type": "object",
        "properties": {
          "adresse1": {
            "type": "string"
          },
          "adresse2": {
            "type": "string"
          },
          "ville": {
            "type": "string"
          },
          "codePostal": {
            "type": "string"
          },
          "pays": {
            "type": "string"
          }
        }
      },
      "ClientDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nom": {
            "type": "string"
          },
          "prenom": {
            "type": "string"
          },
          "adresse": {
            "$ref": "#/components/schemas/Adresse"
          },
          "photo": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "numTel": {
            "type": "string"
          }
        }
      },
      "CommandeClientDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string"
          },
          "dateCommande": {
            "type": "string",
            "format": "date-time"
          },
          "clientDto": {
            "$ref": "#/components/schemas/ClientDto"
          },
          "etat": {
            "type": "string",
            "enum": [
              "EN_PREPARATION",
              "VALIDEE",
              "LIVREE"
            ]
          },
          "commandeLivree": {
            "type": "boolean"
          }
        }
      },
      "CategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string"
          },
          "designation": {
            "type": "string"
          }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "nom": {
            "type": "string"
          },
          "prenom": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "ADMIN",
              "MANAGER"
            ]
          }
        }
      },
      "AuthenticationResponse": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string"
          },
          "refresh_token": {
            "type": "string"
          }
        }
      },
      "AuthenticationRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "ArticleDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "codeArticle": {
            "type": "string"
          },
          "designation": {
            "type": "string"
          },
          "prixUnitaireHt": {
            "type": "number"
          },
          "tauxTval": {
            "type": "number"
          },
          "prixUnitaireTtc": {
            "type": "number"
          },
          "photo": {
            "type": "string"
          },
          "categoryDto": {
            "$ref": "#/components/schemas/CategoryDto"
          }
        }
      },
      "LigneCommandeClientDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "articleDto": {
            "$ref": "#/components/schemas/ArticleDto"
          },
          "commandeClientDto": {
            "$ref": "#/components/schemas/CommandeClientDto"
          }
        }
      },
      "LigneVenteDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "venteDto": {
            "$ref": "#/components/schemas/VenteDto"
          },
          "quantite": {
            "type": "number"
          },
          "prixUnitaire": {
            "type": "number"
          }
        }
      },
      "LigneCommandeFournisseurDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "articleDto": {
            "$ref": "#/components/schemas/ArticleDto"
          },
          "commandeFournisseurDto": {
            "$ref": "#/components/schemas/CommandeFournisseurDto"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Gestion stock auth description",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}